backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic[email]==2.7.1
pyodbc==5.1.0
bcrypt==4.1.3
PyJWT==2.8.0
python-multipart==0.0.9
reportlab==4.2.0
qrcode[pil]==7.4.2
Pillow==10.3.0
twilio==9.1.0
python-dotenv==1.0.1
httpx==0.27.0
opencv-python-headless==4.10.0.84
numpy==2.1.1
requests
pydantic-settings
